Make astype work for `Array` that are sharded. The current behavior is the same as SDA i.e. it round trips via host.
For `Array`'s that are not fully addressable, the current way is to raise an error. Later on I will change the lowering to XLA to go through the pjit lowering path which handles sharded arrays. This will also fix the round trip to host problem for SDAs.

But one thing about running operations on Arrays that are not fully addressable outside of `jit/pjit` is that they will run in SPMD mode i.e. all processes need to execute those ops at the same time. **I am thinking of creating a context manager for that purpose to make it explicit rather than implicit and reduce the possibility of deadlocks. Outside of the context manager, we will raise an error for ops on Arrays that are not fully addressable.**

PiperOrigin-RevId: 457250210